refactor: 💡 (App) Move custom theme into new component
